Adds realtime status updates when building. Handles bundle install errors.

+++ lib/build_runner.rb
@@ -4,9 +4,12 @@ class BuildRunner
     repository = @build.repository
     @build.update_attributes(status: Build::RUNNING)
 
+    @build.send_progress_status(status: Build::RUNNING, message: 'Build Started...')
+
     working_dir = repository.working_directory
     Dir.chdir(working_dir)
-    puts working_dir
+
+    @build.send_progress_status(message: 'Pulling latest changes from github.')
 
     git_reset
     git_pull
@@ -24,28 +27,45 @@ class BuildRunner
       end
     end
 
+    build_status = Build::COMPLETE
+    mutant_json = nil
+    stdout_text = nil
+
     Bundler.with_clean_env do
       ENV['BUNDLE_GEMFILE'] = gemfile
       ENV['BUNDLE_FROZEN'] = '0'
-      bundle_install
-      run_mutant(filter, result_json, stdout_file)
-    end
+      @build.send_progress_status(message: 'Running bundle install...')
+      if bundle_install
+        @build.send_progress_status(message: 'Running Mutant Tests...')
+        run_mutant(filter, result_json, stdout_file)
 
-    mutant_json = nil
-    if File.exist?(result_json) && File.size(result_json) > 0
-      mutant_json = JSON.parse(File.read(result_json))
-      File.unlink(result_json)
-    else
-      log("WARN: JSON output file does not exist.")
+        stdout_text = File.exist?(stdout_file) ? File.read(stdout_file) : nil
+
+        if File.exist?(result_json) && File.size(result_json) > 0
+          mutant_json = JSON.parse(File.read(result_json))
+          File.unlink(result_json)
+        else
+          build_status = Build::ERROR
+          log("WARN: JSON output file does not exist.")
+          @build.send_progress_status(status: Build::ERROR, message: 'Failed running mutant process. See log files.')
+        end
+      else
+        build_status = Build::ERROR
+        @build.send_progress_status(status: Build::ERROR, message: 'Failed running bundle install. See log files.')
+      end
     end
 
     @build.update_attributes({
-      status: mutant_json.is_a?(Hash) ? Build::COMPLETE : Build::ERROR,
+      status: build_status,
       last_sha: current_sha,
       result: mutant_json,
-      stdout: File.read(stdout_file),
+      stdout: stdout_text,
       build_log: File.read(build_log)
     })
+
+    if build_status == Build::COMPLETE
+      @build.send_progress_status(status: Build::COMPLETE, message: 'Build Complete!')
+    end
 
     File.unlink(stdout_file) if File.exist?(stdout_file)
   rescue => e
@@ -74,7 +94,11 @@ class BuildRunner
   def bundle_install
     cmd = %(bundle install --path=vendor)
     log(cmd)
-    log(`#{cmd}`)
+    stdout_str, stderr_str, status = Open3.capture3(cmd)
+    log(stdout_str)
+    log(stderr_str)
+
+    status.success?
   end
 
   def b_exec(cmd)
@@ -109,6 +133,7 @@ class BuildRunner
   end
 
   def run_mutant(filter, result_json, stdout_file)
+    # %(RAILS_ENV=test bundle exec mutant -r ./config/environment --use rspec User)
     cmd = %(bundle exec mutant --include lib/ --json-dump "#{result_json}" --use rspec #{filter.join(' ')} &> "#{stdout_file}")
     log(cmd)
     log(`#{cmd}`)
